调试了一下横向控制器，速度问题依然没有解决

+++ mavros/src/fixed_wing_formation_control/fixd_wing_communicate/test_mavlink.hpp
@@ -25,7 +25,7 @@ private:
 
     serial::Serial sp; //创建一个serial类
 
-    size_t nread;
+    int nread;
 
 public:
     int open_the_seial();
